Administracion de usuarios

+++ backend/src/controllers/usuarios.controller.js
@@ -110,27 +110,46 @@ const hashedPassword = await bcrypt.hash(contrasena, 10);
     }
   },
 
-  deleteUsuario: async (req, res) => {
-    try {
-      const { idUsuario } = req.params;
+ deleteUsuario: async (req, res) => {
+  try {
+    const { idUsuario } = req.params;
 
-      await Usuario.destroy({where: { idUsuario: idUsuario }});
-
-      res.status(200).json({
-        success: true,
-        msg: "usuario eliminado correctamente",
-      });
-
-    } catch (error) {
-      console.error(error);
-      res.status(500).json({
+    // Verificar que el usuario existe
+    const usuario = await Usuario.findByPk(idUsuario);
+    if (!usuario) {
+      return res.status(404).json({
         success: false,
-        msg: process.env.NODE_ENV === "development" //si estas en entorno de desarrollador te muestra el error, si estas del lado de cliente solo te dice que hubo un error interno
-          ? error.message 
-          : "Error interno del servidor",
+        msg: "Usuario no encontrado",
       });
     }
-  },
+
+    // Verificar que no sea un administrador
+    if (usuario.tipoUsuario === 'administrador') {
+      return res.status(403).json({
+        success: false,
+        msg: "No se puede eliminar un administrador",
+      });
+    }
+
+    // Eliminar físicamente el usuario
+    await Usuario.destroy({ where: { idUsuario: idUsuario } });
+
+    res.status(200).json({
+      success: true,
+      msg: "Usuario eliminado correctamente",
+    });
+
+  } catch (error) {
+    console.error(error);
+    res.status(500).json({
+      success: false,
+      msg: process.env.NODE_ENV === "development" 
+        ? error.message 
+        : "Error interno del servidor",
+    });
+  }
+},
+
 
   //Para conectar el frontend con el backend
   loginUsuario: async (req, res) => {
